flipui/requirements.txt
PySide2>=5.14.2.1
requests>=2.25.1
bitarray>=1.2.1
Pillow>=8.1.2
numpy>=1.20.1